First version of code.
Mostly pseudo-code at this point, however, does compile and run.
The purpose of these files is to make sure we are all editing on the same files and to split the code into three sections.

Image rendering file will contain functions used to scan through the view of the robot. 
speed decider file will contain functions used to calculate the speed of the motors

robot.cpp is the file which can be executed and calls these functions.

It will be easiest and cleanest if we can make use of the three files rather then all trying to work on the same file. Please make sure to download all of these files and add them to your AVC_robot folder.